backend/apps/users/serializers.py
from rest_framework import serializers

from .models import User


class CurrentUserSerializer(serializers.ModelSerializer):
    roles = serializers.SlugRelatedField(slug_field="code", many=True, read_only=True)

    class Meta:
        model = User
        fields = (
            "id",
            "username",
            "first_name",
            "last_name",
            "email",
            "phone",
            "status",
            "roles",
        )
